Can drop posts at user location and open them with a tap
