+++ app/src/main/java/com/github/P4rzival/RadiusMessage/MainActivity.java
@@ -47,6 +47,7 @@ import org.osmdroid.config.*;
 import org.osmdroid.tileprovider.tilesource.TileSourceFactory;
 import org.osmdroid.views.MapView;
 import org.osmdroid.views.overlay.Marker;
+import org.osmdroid.views.overlay.Polygon;
 import org.osmdroid.views.overlay.mylocation.GpsMyLocationProvider;
 import org.osmdroid.views.overlay.mylocation.MyLocationNewOverlay;
 
@@ -55,7 +56,7 @@ public class MainActivity extends AppCompatActivity implements TextPostDialog.Te
     public ConstraintLayout parentLayout;
     public PostRenderer postRenderer;
     private Button testPostButton;
-    private Post currentPost;
+    private RadiusPost currentPost;
 
     private float lastTouchX;
     private float lastTouchY;
@@ -107,7 +108,6 @@ public class MainActivity extends AppCompatActivity implements TextPostDialog.Te
                 if(drawData != null && drawData.size() > 0)
                 {
                     addPostToView(drawData.get(drawData.size()-1));
-                    onClickPosts();
                 }
                 Toast.makeText(MainActivity.this
                         , "Post Map Updated"
@@ -140,11 +140,8 @@ public class MainActivity extends AppCompatActivity implements TextPostDialog.Te
         Random rNum = new Random();
         GeoPoint messageLocation = locationNewOverlay.getMyLocation();
         try {
-            //CHNAGE THIS LINE FOR DIFFERENT TEXT!
             test.put("userTextMessage", postText);
             test.put("radius", postRadius);
-            //test.put("locationX", 1 + (1200 - 1) * rNum.nextDouble());
-            //test.put("locationY", 1 + (1300 - 1) * rNum.nextDouble());
             test.put("locationX", messageLocation.getLongitude());
             test.put("locationY", messageLocation.getLatitude());
             test.put("messageDuration", postDuration);
@@ -156,103 +153,12 @@ public class MainActivity extends AppCompatActivity implements TextPostDialog.Te
         postDrawer.createPost(test);
     }
 
-    //Need this in main activity for now
-    public void addPostToView(drawData newPostDrawData){
-        LayoutInflater inflater = (LayoutInflater) getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-        RadiusPost newPost = new RadiusPost(map, newPostDrawData);
-        //Post newPost = new Post(getApplicationContext(), newPostDrawData);
-        //parentLayout.addView(newPost);
-        //postRenderer.postList.add(newPost);
-        //MARKER TEST
-        GeoPoint messageLocation = locationNewOverlay.getMyLocation();
-        /*Drawable nodeIcon = getResources().getDrawable(R.drawable.marker_default);
-        Marker newMarker = new Marker(map);
-        newMarker.setPosition(messageLocation);
-        newMarker.setIcon(nodeIcon);
-        map.getOverlays().add(newMarker);*/
 
+    public void addPostToView(drawData newPostDrawData){
+        RadiusPost newPost = new RadiusPost(map, newPostDrawData,parentLayout);
+        GeoPoint messageLocation = locationNewOverlay.getMyLocation();
         newPost.drawMapPost(messageLocation, map);
         postRenderer.radiusPosts.add(newPost);
-    }
-
-    //For demo 1 keep commented out
-    //Kinda a hack for now, not my favorite way of solving this.
-    //Though it won't go through the whole array, it stops when no input
-    public void onClickPosts(){
-        for (int i = 0; i < postRenderer.postList.size(); i++) {
-            currentPost = postRenderer.postList.get(i);
-            View.OnTouchListener touchListener = new View.OnTouchListener() {
-                @Override
-                public boolean onTouch(View view, MotionEvent motionEvent) {
-                    if (motionEvent.getActionMasked() == MotionEvent.ACTION_DOWN) {
-                        lastTouchX = motionEvent.getX();
-                        lastTouchY = motionEvent.getY();
-                        Post postToOpen = isInRadius(lastTouchX, lastTouchY);
-                        if (postToOpen != null) {
-                            openPostMessage(postToOpen);
-                        }
-                    }
-                    return false;
-                }
-            };
-            currentPost.setOnTouchListener(touchListener);
-        }
-    }
-
-    //Going to use a method like this for our user location to post detection
-    //so might as well do this now
-    //Will need to focus on optimising this in future but for now this is it
-    public Post isInRadius(float touchX, float touchY){
-
-        Post postToOpen = null;
-
-        float distance;
-
-        for (int i = 0; i < postRenderer.postList.size(); i++){
-            drawData postData = postRenderer.postList.get(i).getPostData();
-
-            distance = (float) (Math.pow(touchX - (float) postData.getLocationX(),2) +
-                    Math.pow(touchY - (float) postData.getLocationY(), 2));
-
-            if(Math.sqrt(distance) <= postData.getRadius()){
-                postToOpen = postRenderer.postList.get(i);
-            }
-        }
-        return postToOpen;
-    }
-
-    //Open message after detecting touch
-    public void openPostMessage(Post currentPost){
-
-        LayoutInflater inflater = (LayoutInflater) getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-        View newMessagePopup = inflater.inflate(R.layout.post_message, null);
-
-        int width = 920;
-        int height = 1400;
-        boolean focusable = true;
-
-        final PopupWindow window = new PopupWindow(newMessagePopup, width, height, focusable);
-        window.showAtLocation(parentLayout, Gravity.CENTER, 0,0);
-
-        //Add text to popup from user input
-        TextView currentText = window
-                .getContentView()
-                .findViewById(R.id.messageTextView);
-
-        currentText.setText(currentPost.getPostData().getUserMessageText());
-        newMessagePopup.setOnTouchListener(new View.OnTouchListener() {
-            @Override
-            public boolean onTouch(View v, MotionEvent event) {
-                window.dismiss();
-
-                return true;
-            }
-        });
-
-        Toast.makeText(getApplicationContext()
-                , "Post Opened."
-                , Toast.LENGTH_SHORT).show();
-
     }
 
     //OPEN STREET MAPS FUNCTIONS
